if list is only element, fail test
