+++ src/test/java/com/lperthel/datastructures/PartialReverseLinkedListTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class PartialReverseLinkedListTest {
-	private Solution partialReverseLinkedList = new Solution();
+	private PartiallyReversedLinkedList partialReverseLinkedList = new PartiallyReversedLinkedList();
 	@Test
 	void test_GivenArrayOfNums_ExpectPartiallyReversedPrintedListOfNodes() {
 int left =3;
